Split dual review audit module
