+++ scinoephile/analysis/audit/__init__.py
@@ -5,5 +5,5 @@
 Package hierarchy (modules may import from any above):
 * utils
 * delineation / ocr_fusion / punctuation / review / translation
-* gap_translation / guided_review / guided_translation
+* dual_review / gap_translation / guided_review / guided_translation
 """
