+++ scinoephile/cli/audit/audit_review_dual_cli.py
@@ -8,7 +8,7 @@ from argparse import ArgumentParser
 from collections.abc import Sequence
 from pathlib import Path
 
-from scinoephile.analysis.audit.review import audit_dual_review
+from scinoephile.analysis.audit.dual_review import audit_dual_review
 from scinoephile.analysis.audit.utils import ExtendedAuditFilter
 from scinoephile.cli.helpers.io import read_series
 from scinoephile.common.argument_parsing import (
